Tests and final functionality on gardens

+++ app/assets/javascripts/users/gardens.js
@@ -8,7 +8,7 @@ openFarmApp.controller('gardenCtrl', ['$scope', '$http', 'userService',
 
     $scope.alerts = [];
 
-    $scope.init = function(success, user){
+    $scope.initCallback = function(success, user){
       $scope.currentUser = user;
       $scope.gardens = user.gardens;
       angular.forEach(user.gardens, function(garden){
@@ -20,7 +20,7 @@ openFarmApp.controller('gardenCtrl', ['$scope', '$http', 'userService',
         });
       });
     };
-    userService.getUser(USER_ID, $scope.alerts, $scope.init);
+    userService.getUser(USER_ID, $scope.alerts, $scope.initCallback);
 
     $scope.selectAll = function(garden){
       angular.forEach(garden.garden_crops, function(crop){
@@ -41,6 +41,11 @@ openFarmApp.controller('gardenCtrl', ['$scope', '$http', 'userService',
       }
     };
 
+    $scope.saveGarden = function(garden){
+      gardenService.saveGarden(garden, $scope.alerts);
+      // $scope.saveGardenCropChanges(garden);
+    };
+
     $scope.saveGardenCropChanges = function(garden){
       angular.forEach(garden.garden_crops, function(crop){
         gardenService.saveGardenCrop(garden,
@@ -51,12 +56,14 @@ openFarmApp.controller('gardenCtrl', ['$scope', '$http', 'userService',
 
     $scope.deleteSelected = function(garden){
       angular.forEach(garden.garden_crops, function(crop){
-        if (crop.selected){
+        if (crop.selected && !crop.hide){
           var removeFromList = function(success){
+            console.log(success);
             if (success){
               // don't quite want to deal with splicing arrays
               // across different function calls
               crop.hide = true;
+              console.log(crop);
             }
           };
           gardenService.deleteGardenCrop(garden,
